refactor the locations of the 'noise.dart' part files
