+++ CHANGELOG.md
@@ -1,10 +1,14 @@
+## 2.1.4
+
+- Refactor the locations of the `SimplexNoise` related part files.
+
 ## 2.1.3
 
 - `operator ==` overrides no longer take nullable arguments. This is only
   visible for classes that implement the interfaces defined in this package
   which no longer need to also widen the argument type.
-- Deprecate `SimplexNoise`; see https:github.com/google/vector_math.dart/issues/270
-  for more information.
+- Deprecate `SimplexNoise` (see https:github.com/google/vector_math.dart/issues/270
+  for more information).
 
 ## 2.1.2
