+++ lib/vector_math.dart
@@ -18,7 +18,6 @@ library vector_math;
 import 'dart:math' as math;
 import 'dart:typed_data';
 
-part 'src/vector_math/utilities.dart';
 part 'src/vector_math/aabb2.dart';
 part 'src/vector_math/aabb3.dart';
 part 'src/vector_math/colors.dart';
@@ -29,6 +28,7 @@ part 'src/vector_math/intersection_result.dart';
 part 'src/vector_math/matrix2.dart';
 part 'src/vector_math/matrix3.dart';
 part 'src/vector_math/matrix4.dart';
+part 'src/vector_math/noise.dart';
 part 'src/vector_math/obb3.dart';
 part 'src/vector_math/opengl.dart';
 part 'src/vector_math/plane.dart';
@@ -36,8 +36,8 @@ part 'src/vector_math/quad.dart';
 part 'src/vector_math/quaternion.dart';
 part 'src/vector_math/ray.dart';
 part 'src/vector_math/sphere.dart';
-part 'src/vector_math/third_party/noise.dart';
 part 'src/vector_math/triangle.dart';
+part 'src/vector_math/utilities.dart';
 part 'src/vector_math/vector.dart';
 part 'src/vector_math/vector2.dart';
 part 'src/vector_math/vector3.dart';
