+++ lib/vector_math_64.dart
@@ -18,7 +18,6 @@ library vector_math_64;
 import 'dart:math' as math;
 import 'dart:typed_data';
 
-part 'src/vector_math_64/utilities.dart';
 part 'src/vector_math_64/aabb2.dart';
 part 'src/vector_math_64/aabb3.dart';
 part 'src/vector_math_64/colors.dart';
@@ -29,6 +28,7 @@ part 'src/vector_math_64/intersection_result.dart';
 part 'src/vector_math_64/matrix2.dart';
 part 'src/vector_math_64/matrix3.dart';
 part 'src/vector_math_64/matrix4.dart';
+part 'src/vector_math_64/noise.dart';
 part 'src/vector_math_64/obb3.dart';
 part 'src/vector_math_64/opengl.dart';
 part 'src/vector_math_64/plane.dart';
@@ -36,8 +36,8 @@ part 'src/vector_math_64/quad.dart';
 part 'src/vector_math_64/quaternion.dart';
 part 'src/vector_math_64/ray.dart';
 part 'src/vector_math_64/sphere.dart';
-part 'src/vector_math_64/third_party/noise.dart';
 part 'src/vector_math_64/triangle.dart';
+part 'src/vector_math_64/utilities.dart';
 part 'src/vector_math_64/vector.dart';
 part 'src/vector_math_64/vector2.dart';
 part 'src/vector_math_64/vector3.dart';
