+++ pubspec.yaml
@@ -1,5 +1,5 @@
 name: vector_math
-version: 2.1.3
+version: 2.1.4
 description: A Vector Math library for 2D and 3D applications.
 repository: https://github.com/google/vector_math.dart
